feat: add landing page with terminal aesthetic
Implement unauthenticated landing page with brutalist terminal design.

## Changes

**Infrastructure:**
- Make `/` public route in middleware (auth check server-side)
- Extract dashboard to separate component for code splitting
- Server-side conditional render (landing vs dashboard)
- Dynamic import for dashboard (65% bundle reduction for visitors)

**Landing Page:**
- Hero section with terminal prompt aesthetic (> TRACK. LIFT. GROW._)
- Blinking cursor animation (1s period)
- 3 value prop cards: FAST, SIMPLE, YOURS
- Reuses TerminalPanel, CornerBracket components
- Mobile-first responsive design (stacks → 3-column grid)
- Full dark/light mode support

**Auth Pages:**
- Sign-in page with Clerk component (catch-all route)
- Sign-up page with Clerk component (catch-all route)
- Centered, responsive layout

**Navigation:**
- Conditional auth buttons when unauthenticated
- LOG IN + SIGN UP buttons match terminal aesthetic
- UserButton shown when authenticated

**User Flow:**
1. Visit `/` unauthenticated → landing page
2. Click GET STARTED/SIGN UP → `/sign-up`
3. Complete sign-up → redirect to `/` (dashboard)
4. Sign out → redirect to `/` (landing)
