+++ src/app/sign-in/[[...sign-in]]/page.tsx
@@ -1,0 +1,9 @@
+import { SignIn } from "@clerk/nextjs";
+
+export default function SignInPage() {
+  return (
+    <main className="min-h-screen flex items-center justify-center p-4">
+      <SignIn />
+    </main>
+  );
+}
