+++ src/app/sign-up/[[...sign-up]]/page.tsx
@@ -1,0 +1,9 @@
+import { SignUp } from "@clerk/nextjs";
+
+export default function SignUpPage() {
+  return (
+    <main className="min-h-screen flex items-center justify-center p-4">
+      <SignUp />
+    </main>
+  );
+}
