+++ src/components/landing/BlinkingCursor.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+export function BlinkingCursor() {
+  return (
+    <span className="animate-blink-cursor" aria-hidden="true">
+      _
+    </span>
+  );
+}
